+++ packages/fixie-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fixieai/fixie-common
 
+## 1.0.8
+
+### Patch Changes
+
+- Fix agent deployment.
+
+## 1.0.7
+
+### Patch Changes
+
+- Fix createRevision and agent deployment.
+
 ## 1.0.6
 
 ### Patch Changes